+++ R/utils-cli.R
@@ -1,14 +1,24 @@
 # https://github.com/r-lib/cli/issues/228 ---------------------------------
 
-cli_menu <- function(prompt, not_interactive, choices, quit = integer(), .envir = caller_env()) {
+cli_menu <- function(header,
+                     prompt,
+                     choices,
+                     not_interactive = choices,
+                     quit = integer(),
+                     .envir = caller_env(),
+                     error_call = caller_env()) {
   if (!is_interactive()) {
-    cli::cli_abort(c(prompt, not_interactive), .envir = .envir)
+    cli::cli_abort(
+      c(header, not_interactive),
+      .envir = .envir,
+      call = error_call
+    )
   }
   choices <- sapply(choices, cli::format_inline, .envir = .envir, USE.NAMES = FALSE)
 
-  choices <- paste0(seq_along(choices), " ", choices)
+  choices <- paste0(cli::style_bold(seq_along(choices)), ": ", choices)
   cli::cli_inform(
-    c(prompt, "What do you want to do?", choices),
+    c(header, prompt, choices),
     .envir = .envir
   )
 
@@ -17,12 +27,20 @@ cli_menu <- function(prompt, not_interactive, choices, quit = integer(), .envir 
     if (selected %in% c("0", seq_along(choices))) {
       break
     }
-    cli::cli_inform("Enter an item from the menu, or 0 to exit")
+    cli::cli_inform(
+      "Type a number between 1 and {length(choices)}, or type 0 to exit."
+    )
   }
 
   selected <- as.integer(selected)
   if (selected %in% c(0, quit)) {
-    cli::cli_abort("Quiting...", call = NULL)
+    if (is_testing()) {
+      cli::cli_abort("Quiting...", call = NULL)
+    } else {
+      cli::cli_alert_danger("Quiting...")
+      # simulate user pressing Ctrl + C
+      invokeRestart("abort", cnd)
+    }
   }
   selected
 }
@@ -36,6 +54,10 @@ cli_readline <- function(prompt) {
     options(cli_prompt = testing[-1])
     selected
   } else {
-    readline("Selection: ")
+    readline(prompt)
   }
+}
+
+is_testing <- function() {
+  identical(Sys.getenv("TESTTHAT"), "true")
 }