+++ tests/testthat/test-account-find.R
@@ -21,7 +21,7 @@ test_that("error if no matching account", {
   })
 })
 
-test_that("error if ambiguous accounts", {
+test_that("error if ambiguous accounts in non-interactive environment", {
   mockr::local_mock(accounts = fakeAccounts(c("a", "a", "b"), c("x", "y", "y")))
 
   expect_snapshot(error = TRUE, {
@@ -29,6 +29,19 @@ test_that("error if ambiguous accounts", {
     findAccount("a", NULL)
     findAccount(NULL, "y")
   })
+})
+
+test_that("prompted to pick account in interactive environment", {
+  withr::local_options(
+    rlang_interactive = TRUE,
+    cli_prompt = "2"
+  )
+  mockr::local_mock(accounts = fakeAccounts(c("a", "a", "b"), c("x", "y", "y")))
+
+  expect_snapshot({
+    out <- findAccount()
+  })
+  expect_equal(out, list(name = "a", server = "y"))
 })
 
 test_that("returns account + server when uniquely identified", {