+++ NEWS.md
@@ -1,5 +1,9 @@
 # rsconnect 0.8.30 (development version)
 
+* `deployApp()` now generates an interactive prompt to select 
+  `account`/`server` (if no previous deployments) or 
+  `appName`/`account`/`server` (if multiple previous deployments) (#691). 
+
 * `discoverServer()` has been deprecated; it never worked.
 
 * `deployDoc()` includes a `.Rprofile` in the bundle, if one is found in the